WIP: upgrade actions as script
